major changes
I had to fix the Character Class. there were some feilds and methods that were static. were aren't creating a static character.. so that was fucking everything up. :P after fixing that i had to go in to all the races and classes to fix the calling of those static members to just having a parameter thrown into those methods, which in turn had to change the generator so that it could throw the character its self into the contructors for those races and classes.